Break out AppComponent into seperate file, move out util functions
